requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
streamlit
pandas
numpy
pillow
google-generativeai
scikit-learn
transformers
torch
speechrecognition
dotenv
pillow